+++ gulp/tasks/paris.js
@@ -6,15 +6,11 @@ var runSequence = require('run-sequence');
 
 gulp.task('default', ['harp:serve', 'watch:js']);
 gulp.task('compile', ['harp:compile', 'compile:js']);
-gulp.task('build',   ['build:js', 'build:images', 'build:fonts', 'build:css', 'build:jade', 'build:runtime', 'copy:config', 'copy:locales', 'copy:modernizr'], function () {
+gulp.task('build',   ['build:js', 'build:images', 'build:fonts', 'build:css', 'build:css-exp', 'build:jade', 'build:runtime', 'copy:config', 'copy:locales', 'copy:modernizr'], function () {
 
-    if(fs.existsSync(config.tmp.output + '/client.tpl.js')){
-      fs.renameSync(config.tmp.output + '/client.tpl.js', config.build.assets.javascript + '/client.tpl.js');
-    }
-
-    if(fs.existsSync(config.tmp.output + '/runtime.tpl.js')){
-      fs.renameSync(config.tmp.output + '/runtime.tpl.js', config.build.assets.javascript + '/runtime.tpl.js');
-    }
+  if(fs.existsSync(config.tmp.output + '/runtime.tpl.js')){
+    fs.renameSync(config.tmp.output + '/runtime.tpl.js', config.build.assets.javascript + '/runtime.tpl.js');
+  }
 
 });