+++ src/components/jecoute/script.js
@@ -13,56 +13,175 @@ Paris.jecoute = (function(){
   function jecoute(selector, userOptions){
     var $el = $(selector),
       options = $.extend({}, defaultOptions, userOptions),
-      $form,
-      $formElements;
+      $message, $form, $formElements, $upload, $output,
+      files = [];
 
-    function init(){
+    function init() {
       initOptions();
 
+      $message = $el.find('.jecoute-message');
       $form = $el.find('.jecoute-form');
       $formElements = $form.find('input, textarea, button');
+      $upload = $formElements.filter('input[type="file"]');
+      $output = $form.find('output');
 
       $form.on('submit', onSubmitForm);
 
       if (Cookies.get(Paris.config.cookies.email.name)) {
         $formElements.filter('input[name="email"]').attr('value', Cookies.get(Paris.config.cookies.email.name));
       }
+
+      if ($upload.length) {
+        if (window.File && window.FileReader && window.FileList) {
+          $upload.on('change', onFileUpload);
+        }
+        else {
+          console.log('The File APIs are not fully supported in this browser.');
+        }
+      }
     }
 
     function initOptions() {
-      $.each($el.data(), function(key, value){
+      $.each($el.data(), function(key, value) {
         options[key] = value;
       });
+    }
+
+    function removeUploadedFile(e) {
+      var $li = $(this).parent();
+      var index = $li.index();
+
+      files.splice(index, 1);
+      $li.remove();
+
+      isBiggerSize();
+    }
+
+    function onFileUpload(e) {
+
+      for (var i = 0, f; f = e.target.files[i]; i++) {
+        var html = '';
+        html += '<li class="output-item">';
+        // html += f.name;
+        html += (f.name.length > 40) ? f.name.substr(0, 19) + ' ... ' + f.name.substr(f.name.length-18, f.name.length) : f.name;
+        html += '<i>(' + humanSize(f.size) + ')</i>';
+        html += '<span class="output-item-remove icon-close-rounded"></span>';
+        html += '</li>';
+
+        $output.find('.output-items').append(html);
+        files.push(f);
+      }
+
+      $output.find('.output-item-remove').on('click', removeUploadedFile);
+      isBiggerSize();
     }
 
     function onSubmitForm(e) {
       e.preventDefault();
 
-      // Serialize array before disabling form elements
-      var data = $(this).serializeArray();
-      $formElements.attr('disabled', 'disabled');
+      var $policy = $formElements.filter('input[type="checkbox"]');
+      if ($policy.length && !$policy.is(':checked')) {
+        showError(Paris.i18n.t("jecoute/error/policy"));
+        return;
+      }
 
-      // Insert current url
-      data.push({
-        name:   'url',
-        value:  window.location.href
-      });
+      if (isBiggerSize()) {
+        showError(Paris.i18n.t("jecoute/error/file_size"));
+        return;
+      }
+
+      var data;
+      if (files.length) {
+        data = new FormData();
+
+        var $textarea = $formElements.filter('textarea');
+        data.append($textarea.attr('name'), $textarea.val());
+
+        var $email = $formElements.filter('input[type="email"]');
+        data.append($email.attr('name'), $email.val());
+
+        if ($policy.length) {
+          data.append($policy.attr('name'), $policy.val());
+        }
+
+        data.append('url', window.location.href);
+
+        // upload files at the end for Sails not to trunc data
+        for (var i = 0, f; f = files[i]; i++) {
+          data.append('files', f, f.name);
+        }
+      }
+      else {
+        data = $(this).serializeArray();
+
+        data.push({
+          name: 'url',
+          value: window.location.href
+        });
+      }
+
+      $formElements.attr('disabled', 'disabled');
 
       saveQuestion(data);
     }
 
     function saveQuestion(data) {
-      $.ajax({
-        url:      $form.attr('action'),
-        type:     $form.attr('method'),
-        data:     data,
+      var ajaxOptions = {
+        url: $form.attr('action'),
+        type: $form.attr('method'),
+        data: data,
         complete: onQuestionSaved
-      });
+      };
+
+      if (files.length) {
+        ajaxOptions.contentType = false;
+        ajaxOptions.processData = false;
+      }
+
+      $.ajax(ajaxOptions);
     }
 
-    function onQuestionSaved() {
-      $form.hide().after('<p>'+ options.thanks +'</p>');
+    function onQuestionSaved(jqXHR, status) {
+      $form.hide();
+      $message.text(options.thanks).show();
     }
+
+    /////////////
+    // HELPERS //
+    /////////////
+
+    function showError(msg) {
+      $message.text(msg).addClass('text-error').show();
+      window.setTimeout(function() {
+        $message.show().removeClass('text-error').text('');
+      }, 5000);
+    }
+
+    function isBiggerSize(maxSize) {
+      var size = 0;
+      var ms = maxSize || 10*1048576; // 10Mo
+
+      for (var i = 0, f; f = files[i]; i++) {
+        size += f.size;
+      }
+
+      var isBigger = size > ms;
+      $form.find('.upload-size').toggleClass('text-error', isBigger);
+
+      return isBigger;
+    }
+
+    function humanSize(bytes) {
+      var sizes = ['octets', 'Ko', 'Mo', 'Go', 'To'];
+      if (bytes === 0) return '0 octet';
+
+      var i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)));
+      return Math.round(bytes / Math.pow(1024, i), 2) + ' ' + sizes[i];
+    }
+
+    //////////
+    // INIT //
+    //////////
 
     init();