+++ src/javascript/main.js
@@ -7,9 +7,12 @@ require('./lib/i18n');
 require('./lib/requestanimationframe');
 require('./lib/responsive');
 require('./lib/scroll');
+require('./lib/url');
 
 // load our components
 require('../components/accordion/script');
+require('../components/form/script');
+require('../components/gallery/script');
 require('../components/html/script');
 require('../components/jecoute/script');
 require('../components/news-push/script');
@@ -25,6 +28,7 @@ require('../modules/block-content-newsletter/script');
 require('../modules/button-top/script');
 require('../modules/gallery-ugc/script');
 require('../modules/notice/script');
+require('../modules/person-block/script');
 require('../modules/poll/script');
 require('../modules/quick-access/script');
 require('../modules/rheader/script');
@@ -43,5 +47,5 @@ require('./lib/cnil');
 require('./lib/egg');
 
 var Paris = window.Paris || {};
-Paris.version = "1.0.7";
-Paris.templates = {};
+Paris.version = "1.2.0";
+Paris.templates = require('./client.tpl');