+++ src/templates/list-persons/script.js
@@ -5,29 +5,43 @@ var jade = require('jade');
 var values = require('lodash.values');
 var map = require('lodash.map');
 var sortBy = require('lodash.sortby');
+var forEach = require('lodash.foreach');
+var map = require('lodash.map');
 var algoliasearch = require('algoliasearch');
+var algoliaSearchHelper = require('algoliasearch-helper');
 
 var Paris = window.Paris || {};
 
 Paris.listPersons = (function(){
 
   var defaultOptions = {
-    index: 'persons', // the Algolia index to use (should be defined in config.js)
-    fields: { // matching the names of Algolia fields
-      link: 'url',              // the field to use as a link
-      title: 'prenom_nom',          // the field to use as the person-block title
-      text: 'mandat',            // the field to use as the person-block text
-      image: 'portrait',        // the field to use as image
-      group: 'groupe_politique' // the field to use in the person-block text
+    // the Algolia index to use (should be defined in config.js)
+    index: 'persons',
+    // matching the names of Algolia fields
+    fields: {
+      // the field to use as a link
+      link: 'url',
+      // the field to use as the person-block title
+      title: 'prenom_nom',
+      // the field to use as the person-block text
+      text: 'mandat',
+      // the field to use as image
+      image: 'portrait',
+      // the field to use in the person-block text
+      group: 'groupe_politique'
     },
-    resultsPerPage: 8, // the number of results to display per page
-    facets: ["groupe_politique", "secteur"], // the available facets that will be displayed in the left column (should have been created on Algolia)
-                       // you can set the name displayed in the left column in locales.js (key: $LOCALE/search_results/facets/$YOUR_FACET)
-    addFacetFilter: null // the facetFilter you want to always add by default (useful for filtering results)
+    // the number of results to display per page
+    resultsPerPage: 100,
+    // the available facets that will be displayed in the left column (should have been created on Algolia)
+    // you can set the name displayed in the left column in locales.js (key: $LOCALE/search_results/facets/$YOUR_FACET)
+    facets: ["groupe_politique", "secteur"],
+    // the facetFilter you want to always add by default (useful for filtering results)
+    addFacetFilterJson: null,
+    algoliaHelperParams: {facets: ['mandat'], disjunctiveFacets: ['groupe_politique', 'secteur']}
   };
 
-  function listPersons(selector, userOptions){
-    var $el     = $(selector),
+  function listPersons(selector, userOptions) {
+    var $el = $(selector),
       options = $.extend({}, defaultOptions, userOptions),
       api = {},
       $searchFieldInput,
@@ -37,13 +51,26 @@ Paris.listPersons = (function(){
       currentFacets = [],
       algolia,
       index,
-      currentPage = 0;
+      helper,
+      currentPage = 0,
+      isFiltered;
 
-    function init(){
+    function init() {
       initOptions();
 
       algolia = algoliasearch(Paris.config.algolia.id, Paris.config.algolia.api_key);
       index = algolia.initIndex(Paris.config.algolia.indexes[options.index]);
+      helper = algoliaSearchHelper(algolia, Paris.config.algolia.indexes[options.index], options.algoliaHelperParams);
+
+      // Add mandatory facet filters
+      if (options.addFacetFilterJson) {
+        $.each(options.addFacetFilterJson, function(index, facetFilter) {
+          helper.addRefine(facetFilter.facetName, facetFilter.facetValue);
+        });
+      }
+
+      helper.on('result', onSearchResults);
+      // helper.on('error', onSearchError);
 
       $searchFieldInput = $el.find('#search-person');
       $results = $el.find('#results');
@@ -51,27 +78,29 @@ Paris.listPersons = (function(){
       $facetsContainer = $el.find('#facets');
 
       $searchFieldInput.on('input', onInput);
-      $facetsContainer.on('change', 'input[type=checkbox]', updateFacets);
+
+      $facetsContainer.on('click', 'a', toggleFacet);
+      $facetsContainer.on('change', 'select', updateFacets);
       $pagination.on('click', 'a', onClickPagination);
 
-      // If the search field is not empty, trigger the search on page load
-      if ($searchFieldInput.val() != "") {
-        $searchFieldInput.trigger('input');
+      // If the search field is not empty or if there is more than one facet selected,
+      // update facets and launch search
+      if ($searchFieldInput.val() != "" || _moreThanOneFacet()) {
+        updateFacets();
       }
 
       $el.data('api', api);
     }
 
     function initOptions() {
-      $.each($el.data(), function(key, value){
+      $.each($el.data(), function(key, value) {
         options[key] = value;
       });
     }
 
     function onInput() {
       currentPage = 0;
-      var query = $searchFieldInput.val();
-      launchSearch(query);
+      launchSearch();
     }
 
     function scrollToResultsTop() {
@@ -82,46 +111,39 @@ Paris.listPersons = (function(){
       });
     }
 
-    function launchSearch(query){
-      if (typeof query === 'undefined') {
-        query = $searchFieldInput.val();
+    function launchSearch() {
+      var query = $searchFieldInput.val();
+
+      isFiltered = query ? true : false;
+      helper.setQuery(query);
+      helper.setCurrentPage(currentPage);
+
+      // add selected facets to algolia search
+      var helperState = JSON.parse(JSON.stringify(helper.getState()));
+      $.each(options.facets, function(index, facet) {
+        if (currentFacets[facet] && currentFacets[facet].length) {
+          isFiltered = true;
+        }
+        helperState.disjunctiveFacetsRefinements[facet] = currentFacets[facet];
+      });
+      helper.setState(helperState);
+
+      helper.setQueryParameter('hitsPerPage', isFiltered ? options.resultsPerPage : 8);
+
+      // set url query parameters
+      var urlQuery;
+      if (isFiltered) {
+        urlQuery = $.extend({}, {search: query}, currentFacets);
       }
-
-      var params = {
-        // Pagination params
-        page: currentPage,
-        hitsPerPage: options.resultsPerPage,
-
-        // Explicitly request necessary facets (as defined in options)
-        facets: options.facets.join(','),
-        facetFilters: [],
-
-        // Explicitly request necessary attributes (as defined in options)
-        attributesToRetrieve: values(options.fields).join(',')
-      };
-
-      // If some facets filters are active, add them to the request
-      if (currentFacets.length !== 0) {
-        params.facetFilters = currentFacets;
+      else {
+        urlQuery = (currentPage > 0) ? {page: currentPage + 1} : {};
       }
+      Paris.url.setQueryString(urlQuery);
 
-      // Add mandatory facet filters
-      if (options.addFacetFilter !== null) {
-        params.facetFilters.push(options.addFacetFilter);
-      }
-
-      // Launch the search using Algolia
-      index.search(query, params, onSearchResults);
+      helper.search();
     }
 
-    function onSearchResults(err, data) {
-      if (err) {return;}
-
-      if (data.query !== $searchFieldInput.val()) {
-        // do not take out-dated answers into account
-        return;
-      }
-
+    function onSearchResults(data) {
       renderResults(data);
       renderFacets(data);
     }
@@ -155,7 +177,7 @@ Paris.listPersons = (function(){
           person_block_data.text.push(hit[options.fields.group]);
           person_block_data.text = person_block_data.text.join(", ");
 
-          var person_block = Paris.templates.templatizer["person-block"]["person-block"](person_block_data);
+          var person_block = Paris.templates["person-block"]["person-block"](person_block_data);
 
           $results.append(person_block);
         });
@@ -164,26 +186,28 @@ Paris.listPersons = (function(){
       renderPagination(data);
     }
 
-    function renderPagination(data){
-      if (data.nbPages > 1) {
+    function renderPagination(data) {
+      if ((data.nbPages > 1) || !isFiltered) {
         var pagination_data = {
           "text": {
+            "label": Paris.i18n.t("pagination/label"),
             "prev": Paris.i18n.t("pagination/prev"),
-            "next": Paris.i18n.t("pagination/next")
+            "next": Paris.i18n.t("pagination/next"),
+            "link_title": Paris.i18n.t("pagination/link_title")
           },
-          "base_url": "#",
-          "url": "#page-${page}",
+          "base_url": window.location.pathname,
+          "url": "?page=${page}",
           "current": data.page + 1,
           "total": data.nbPages
         };
-        var pagination = Paris.templates.templatizer["pagination"]["pagination"](pagination_data);
+        var pagination = Paris.templates["pagination"]["pagination"](pagination_data);
         $pagination.html(pagination);
       } else {
         $pagination.empty();
       }
     }
 
-    function onClickPagination(e){
+    function onClickPagination(e) {
       e.preventDefault();
       var page = $(this).data('page');
       if (page === "prev") {
@@ -198,81 +222,105 @@ Paris.listPersons = (function(){
     }
 
     function renderFacets(data) {
+      var allFacets = {};
+      $.each(['disjunctiveFacets', 'facets'], function(index, facetType) {
+        $.each(data[facetType], function(index, facet) {
+          allFacets[facet.name] = facet.data;
+        });
+      });
 
-      if (currentFacets.length === 0
-        && !$.isEmptyObject(options.facets)
-        && !$.isEmptyObject(data.facets)
-      ) {
-        $facetsContainer.empty();
+      $facetsContainer.empty();
 
-        $.each(options.facets, function(index, facet) {
+      $.each(options.facets, function(index, facet) {
 
-          var block_aside_checkboxes_data = {
-            title: Paris.i18n.t("search_results/facets/"+facet),
-            name: facet,
-            modifiers: facet === "secteur" ? ["two-cols"] : [],
-            items: []
-          };
+        var block_aside_checkboxes_data = {
+          title: Paris.i18n.t("search_results/facets/"+facet),
+          name: facet,
+          modifiers: facet === "secteur" ? ["two-cols"] : [],
+          items: []
+        };
 
-          if (facet === "secteur") {
-
-            var facetArray = map(data.facets[facet], function(number, name){ return {
+        if (facet === "secteur") {
+          var facetArray = map(allFacets[facet], function(number, name) {
+            return {
               value: name,
               text: name.replace(" arrondissement", ""),
-              number: Paris.i18n.formatNumber(number)
-            }; });
-            block_aside_checkboxes_data.items = sortBy(facetArray, function(o){
-              return parseInt(o.value, 10);
+              number: Paris.i18n.formatNumber(number),
+              checked: $.inArray(name, currentFacets[facet]) !== -1
+            };
+          });
+
+          block_aside_checkboxes_data.items = sortBy(facetArray, function(o) {
+            return parseInt(o.value, 10);
+          });
+        }
+        else if (allFacets[facet]) {
+          $.each(allFacets[facet], function(name, number) {
+            block_aside_checkboxes_data.items.push({
+              value: name,
+              text: name,
+              number: Paris.i18n.formatNumber(number),
+              checked: $.inArray(name, currentFacets[facet]) !== -1
             });
+          });
+        }
 
-          } else {
-
-            $.each(data.facets[facet], function (name, number) {
-              block_aside_checkboxes_data.items.push({
-                value: name,
-                text: name,
-                number: Paris.i18n.formatNumber(number)
-              });
-            });
-
-          }
-
-          var facet_block = Paris.templates.templatizer["block-aside-checkboxes"]["block-aside-checkboxes"](block_aside_checkboxes_data);
-          $facetsContainer.append(facet_block);
-        });
-      }
+        var facet_block = Paris.templates["block-aside-checkboxes"]["block-aside-checkboxes"](block_aside_checkboxes_data);
+        $facetsContainer.append(facet_block);
+      });
     }
 
     function updateFacets() {
-      currentFacets = [];
+      currentFacets = {};
+
       $.each(options.facets, function(index, facet) {
         var facetValues = [];
         var $select = $facetsContainer.find(".block-aside-select[name^='" + facet + "']");
-        var $checkboxes = $facetsContainer.find("input[type=checkbox][name^='" + facet + "']:checked");
+        // var $checkboxes = $facetsContainer.find("input[type=checkbox][name^='" + facet + "']:checked");
+        var $checkboxes = $facetsContainer.find("a[data-name^='" + facet + "'].checked");
 
         if ($select.is(':visible')) {
           if ($select.val() === null) {return;}
-          $.each($select.val(), function(i, value){
-            facetValues.push($select.attr("name").replace("[]", "") + ":" + value);
+          $.each($select.val(), function(i, value) {
+            facetValues.push(value);
           });
         } else {
-          $checkboxes.each(function(){
-            if ($(this).val() === null) {return;}
-            facetValues.push($(this).attr("name").replace("[]", "") + ":" + $(this).val());
+          $checkboxes.each(function() {
+            // var value = $(this).val();
+            var value = $(this).data('value');
+            if (value === null) {return;}
+            facetValues.push(value);
           });
         }
 
         if (facetValues.length > 0) {
-          currentFacets.push(facetValues);
+          currentFacets[facet] = facetValues;
         }
       });
+
       onInput();
     }
 
+    function toggleFacet(event) {
+      if (event) {event.preventDefault();}
+
+      $(this).toggleClass('checked');
+      updateFacets();
+    }
+
+    function _moreThanOneFacet() {
+      var queryObj = Paris.url.parseQueryString();
+      var nbOfFacets = 0;
+
+      $.each(options.facets, function(index, facetName) {
+        nbOfFacets += queryObj[facetName] ? queryObj[facetName].length : 0;
+      });
+
+      return nbOfFacets > 1;
+    }
 
     // The API for external interaction
-
-    api.search = function(query){
+    api.search = function(query) {
       $searchFieldInput.val(query).trigger('input');
     };
 
@@ -281,8 +329,8 @@ Paris.listPersons = (function(){
     return $el;
   }
 
-  return function(selector, userOptions){
-    return $(selector).each(function(){
+  return function(selector, userOptions) {
+    return $(selector).each(function() {
       listPersons(this, userOptions);
     });
   };